状態管理・API通信をカスタムフックに移行
